fixes related to function returning array: do not simplify loads after function calls that may change an array offset during inlining can now be negative

+++ crates/noirc_evaluator/src/ssa/inline.rs
@@ -271,8 +271,7 @@ pub fn inline_in_block(
                     //Compute the new address:
                     //TODO use relative addressing, but that requires a few changes, mainly in acir_gen.rs and integer.rs
                     let b = stack_frame.get_or_default(a);
-                    //n.b. this offset is always positive
-                    let offset = ctx.mem.arrays[b as usize].adr - ctx.mem.arrays[a as usize].adr;
+                    let offset = ctx.mem.arrays[b as usize].adr as i32 - ctx.mem.arrays[a as usize].adr as i32;
                     let index_type = ctx[new_left].get_type();
                     let offset_id =
                         ctx.get_or_create_const(FieldElement::from(offset as i128), index_type);
@@ -291,7 +290,7 @@ pub fn inline_in_block(
                 }
                 Operation::Store(a) => {
                     let b = stack_frame.get_or_default(a);
-                    let offset = ctx.mem.arrays[b as usize].adr - ctx.mem.arrays[a as usize].adr;
+                    let offset = ctx.mem.arrays[b as usize].adr as i32 - ctx.mem.arrays[a as usize].adr as i32;
                     let index_type = ctx[new_right].get_type();
                     let offset_id =
                         ctx.get_or_create_const(FieldElement::from(offset as i128), index_type);
